agents/requirements.txt
langgraph>=0.4.0
langchain>=0.3.20
langchain-groq>=0.3.0
langchain-community>=0.3.20
python-dotenv>=1.0.0
pydantic>=2.10.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.28.0
